# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user account",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              },
              "examples": {
                "student": {
                  "summary": "Register a student account",
                  "value": {
                    "email": "student1@know.mn",
                    "password": "Student!1234",
                    "role": "STUDENT"
                  }
                },
                "instructor": {
                  "summary": "Register an instructor account",
                  "value": {
                    "email": "instructor1@know.mn",
                    "password": "Admin!1234",
                    "role": "INSTRUCTOR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with email and password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "student": {
                  "summary": "Student login",
                  "value": {
                    "email": "student1@know.mn",
                    "password": "Student!1234"
                  }
                },
                "admin": {
                  "summary": "Admin login",
                  "value": {
                    "email": "admin@know.mn",
                    "password": "Admin!1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              },
              "examples": {
                "refresh": {
                  "summary": "Refresh token request",
                  "value": {
                    "refreshToken": "paste-refresh-token-from-login-response"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout current session (revokes current access token)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "summary": "Logout from all sessions (revokes all tokens)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out from all sessions"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current authenticated user profile",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "summary": "Change password (revokes all sessions after change)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              },
              "examples": {
                "changePassword": {
                  "summary": "Change current user password",
                  "value": {
                    "currentPassword": "Student!1234",
                    "newPassword": "NewStudent!1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "401": {
            "description": "Unauthorized or wrong current password"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "operationId": "AuthController_listUsers",
        "summary": "[Admin] List all users with pagination",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users listed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/auth/users/{id}/status": {
      "patch": {
        "operationId": "AuthController_updateUserStatus",
        "summary": "[Admin] Activate or deactivate a user account",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Target user UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              },
              "examples": {
                "deactivate": {
                  "summary": "Deactivate user",
                  "value": {
                    "isActive": false
                  }
                },
                "activate": {
                  "summary": "Activate user",
                  "value": {
                    "isActive": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "info": {
    "title": "LMS Auth API",
    "description": "Authentication, session, and user administration endpoints",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication and current-user endpoints"
    },
    {
      "name": "Users",
      "description": "Admin user management endpoints"
    },
    {
      "name": "Health",
      "description": "Service health checks"
    }
  ],
  "servers": [
    {
      "url": "/api",
      "description": "Gateway or service API prefix"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the accessToken returned by /auth/login."
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "SUPER_ADMIN",
          "ADMIN",
          "INSTRUCTOR",
          "STUDENT"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "student1@know.mn",
            "description": "Unique email address used for login."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Student!1234",
            "description": "At least 8 chars with uppercase, lowercase, number, and special char"
          },
          "role": {
            "default": "STUDENT",
            "description": "Optional role. Public registration normally creates STUDENT accounts.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "student1@know.mn",
            "description": "Email address of the user account."
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "Student!1234",
            "description": "Plain text password. It is validated by the auth service and never returned."
          }
        }
      },
      "RefreshTokenDto": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Refresh token issued by login or token refresh."
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "Student!1234",
            "description": "Current password for the authenticated user."
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "example": "NewStudent!1234",
            "description": "At least 8 chars with uppercase, lowercase, number, and special char"
          }
        }
      },
      "UserQueryDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "minimum": 1,
            "default": 1,
            "example": 1,
            "description": "Page number for paginated user results."
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "default": 20,
            "example": 20,
            "description": "Number of users returned per page."
          },
          "role": {
            "example": "STUDENT",
            "description": "Optional role filter for admin user search.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        }
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "required": [
          "isActive"
        ],
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Set to false to deactivate the user, or true to reactivate the user."
          }
        }
      }
    }
  }
}